Move variable app queries to its own folder
