+++ src/controllers/Admin-Web-App/appQueries/variables.js
@@ -1,4 +1,6 @@
-const prettyFormatDate = require('./prettyFormaters')
+
+const { QueryVariablesByUser } = require('../../../helpers/common-queries/variable-queries')
+const prettyFormatDate = require('../../../helpers/data-cleaners-react-app/prettyFormaters')
 
 function removeSensitiveVariableData (variables) {
     let cleanedData = {}
@@ -29,4 +31,19 @@ function removeSensitiveVariableData (variables) {
     return cleanedData
 }
 
-module.exports = removeSensitiveVariableData
+async function getVariables (req, res) {
+    try {
+        let { user } = req 
+        let userVariables = await QueryVariablesByUser(user)
+        let cleanedData = removeSensitiveVariableData(userVariables)
+        res.json(cleanedData)
+    } catch (error) {
+        console.error(error)
+        res.sendStatus(500)
+    }
+} 
+
+
+module.exports = {
+    getVariables, 
+}
